Export the trained model
When the model is trained, in order to run an inference service to serve
it, the model should be exported. Two optional parameters are
introduced:
  "-save NAME"
  "-save_version VERSION"
By default, the model is not exported. If "-save NAME" is specified, the
model is saved using given NAME. If "-save_version VERSION" is
specified, together with "-save NAME", the model is saved using given
NAME and VERSION. The "-save_version" is ignored, if "-save" is missing.
By default, version "001" is used. Models are exported in directory:
 models/<NAME>-<outcome>/<VERSION>/
and are compressed in file:
 models/<NAME>-<outcome>/<NAME>-<outcome>-<VERSION>.tar.gz
The exported models are tested with kserve, the layout of directories and
archive file is designed in a way kserve tensorflow predictor expects.

fixes #2

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>