Improve project structure
Renamed the main source code path from ml-conversational-analytic-tool
to mcat in order to make it a package. Long package names are discouraged
and dashes are invalid symbols (see PEP8).

Moved tests out of the mcat package and made it a package too. Updated
import paths.

Updated the auto-generated docs and the README.

Signed-off-by: Teodora Sechkova <tsechkova@vmware.com>